use the keywords api for freefem
