+++ .rubocop_todo.yml
@@ -18,7 +18,6 @@ Lint/EmptyBlock:
 # SupportedStylesForLeadingUnderscores: disallowed, required, optional
 Naming/MemoizedInstanceVariableName:
   Exclude:
-    - 'lib/rouge/lexers/freefem.rb'
     - 'lib/rouge/lexers/verilog.rb'
 
 # Offense count: 16
@@ -36,7 +35,6 @@ Naming/MethodName:
 Naming/VariableName:
   Exclude:
     - 'lib/rouge/lexers/dafny.rb'
-    - 'lib/rouge/lexers/freefem.rb'
     - 'lib/rouge/lexers/igorpro.rb'
     - 'lib/rouge/lexers/kotlin.rb'
     - 'lib/rouge/lexers/xpath.rb'
@@ -73,7 +71,6 @@ Rouge/NoBuildingAlternationPatternInRegexp:
 # Offense count: 20
 Rouge/NoHugeCollections:
   Exclude:
-    - 'lib/rouge/lexers/freefem.rb'
     - 'lib/rouge/lexers/hql.rb'
     - 'lib/rouge/lexers/idlang.rb'
     - 'lib/rouge/lexers/igorpro.rb'
